requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers==4.40.0
torch==2.2.0
spacy==3.6.0
pymongo==4.4.0
langdetect==1.0.9
deep-translator>=1.9.3,<1.10.0
python-dotenv==1.0.0
rapidfuzz==2.16.1
pytest==7.4.0